requirements.txt
gunicorn==19.9.0
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
numpy==1.19.0
MarkupSafe==1.1.1
Pillow==7.2.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
